FrameData: 	- changed interpolation type for IPeerAssignmentToSlot to long to avoid 	interpolating peer slot assignment
Shifted a lot of files into new folders to sepperate out the networking
library from the per project class files

GamePacketFacrory:
	- commented out code with plans to remove in the future

Packet:
	- commented out code with plans to remove in the future

SimManager:
	- commented out code with plans to remove in the future

ProjectSim: project for the game specific simulation code

ProcessPeer*:
	- moved UserConnectionChange and MessagePayloadWrapper
	into sharedTypes project
	- moved tick rate settings into sim settings class

ISimTickRateSettings:
	- firs pass for interface for settings to store into on sim tick
	and time settings

SimProcessorSettings:
	- Implemented ISimTickRateSettings

SimProcessorSettingsInterface:
	- added code to convert unity settings interface into project
	settings for time settings code

ActiveGameManager:
	- added interpolation settings
	- added frame data interpolation class

ActiveGameManagerSceneTeseter:
	- added interpolation error correction settings
	- added debug draw component to draw interpolated sim state

TestingSimManager:
	- changed from using network data bridge to using a generic
	ISimTimeProvider
	- Changed code to implment IsimFrameDataProvider
	- added extra constraints to FrameData and Settings
	- removed tick rate values (Moves into settings)

LocalPeerInputManager:
	- moved UserInputGLobalMessage into a new SharedTypesProject


FrameDataInterpolatorGen:
	- moved code into a game specific project for interpolation
	- cleaned up code and made it into a self contained type
	- Added Implementation of IFrameDataInterpolator for
	type constraining in generic types

InterpolatedFrameDataGen:
	- added implementation of IInterpolatedFrameData
	- added match format function so the interpolation data
	has the same size arrays as a target frame data object

IFrameData:
	- added interface to use as type constraint when bulding
	generic classes

IGameTimeProvider:
	- interface for difining class that can provide the
	datetime object used when working out what tick the sim
	should be up to

ISimFrameDataProvider:
	- interface for classes that can provide 2 frame data objects
	for the frame data interpolator to interpolate between

FrameDataInterpolationManager:
	- first pass at an interpolation manager that is not sim specific
	that can be used to interpolate frame data from a generic source

+++ Assets/Code/CoreGameSim/FrameData.cs
@@ -19,7 +19,7 @@ namespace Sim
             }
         }
                
-        [FrameDataInterpilationTypeAttribute(typeof(float), FrameDataInterpilationTypeAttribute.InterpolationType.None)]
+        [FrameDataInterpilationTypeAttribute(typeof(long), FrameDataInterpilationTypeAttribute.InterpolationType.None)]
         public long[] m_lPeersAssignedToSlot;
 
         #region IPeerSlotAssignmentFrameData
